step 0

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ndarray/BaseNDArray.java
@@ -2087,7 +2087,7 @@ public abstract class BaseNDArray implements INDArray, Iterable {
 
     @Override
     public INDArray subArray(ShapeOffsetResolution resolution) {
-        int[] offsets = resolution.getOffsets();
+        long[] offsets = resolution.getOffsets();
         int[] shape = resolution.getShapes();
         int[] stride = resolution.getStrides();
